fix lab calculation error

+++ LivingLab.Core/Entities/DTO/EnergyUsageDTOs/LabEnergyUsageDTO.cs
@@ -6,7 +6,7 @@ public class LabEnergyUsageDTO
 {
     public string? LabLocation { get; set; }
 
-    public int TotalEnergyUsage { get; set;}
+    public double TotalEnergyUsage { get; set;}
 
     public double EnergyUsageCost { get; set;}
